Fix test
sort_points_by_argument.test.cpp が通らないバグを修正する．

+++ test/yosupo/geometry/sort_points_by_argument.test.cpp
@@ -12,7 +12,7 @@ int main(void) {
     for (int i = 0; i < n; ++i) {
         int x, y;
         std::cin >> x >> y;
-        v.emplace_back(x, y, std::atan2(y, x));
+        v.emplace_back(x, y, std::atan2((long double)y, x));
     }
     std::sort(v.begin(), v.end(), [&](auto x, auto y) { return get<2>(x) < get<2>(y); });
     for (auto &&[x, y, z] : v) std::cout << x << ' ' << y << '\n';
